py/requirements.txt
wheel
ifaddr
lxml
uvicorn[standard]
fastapi
paho-mqtt

